Add Edition#metadata method
Provides structured metadata (text and hrefs) of multiple types for the
edition.

Override this method in any class or mixin which needs to provide
metadata for the edition. The method is expected to always return a
hash.

In app/models/publication.rb there is an example of metadata being
removed rather than added.

Also adds a with_url scope to NationalApplicability.