Update friendly_id config
Note that this version of friendly_id has a couple of differences to
the previous version:

1) The default sequence seperator used when there is a slug clash is
now a single '-', rather than a double-dash that was used previously.
I've left the config to use the new single-dash, but this can be set
back to double-dash if we deem this necessary.

2) Previous versions of friendly_id appended a numeric sequence to
make slugs unique. The numeric sequence has now been replaced with a
UUID (to simplify the code and make it more thread-safe).

This second change results in very ugly slugs. However, friendly_id
provides a new mechanism to reduce the chance of such clashes by
allowing you to specify slug candidates (see
lib/friendly_id/slugged.rb in the friendly_id gem). I'll explore using
this in a later commit.