Update localised routing patch
The routing patch breaks on Rails 4. I've updated it to behave as it
did before, but unfortunately this doesn't quite do the trick and
localised routing now only works for non-resourceful routes.

I've tracked the problem down to a change in the code that builds the
routes for a restful resources, specifically the scope and
apple_common_behaviours_for methods. These behave slightly differently
in Rails 4 and now the routing code incorrectly decides that a scope
is being applied and fails to pass in the :localised option to the
resultant route-building code.

I didn't get far enough to come up with a suitable workaround that
doesn't involve some horrid monkey-patching of the routing code. An
alternative solution might be to write a new routing methods, rather
than further patching the existing ones, e.g. something like a
"localised_resources" method.