Disable the text extraction feature.
Since we removed the initialiser override from our deployment code which
disables this feature in production, this will default to enabling the
feature, which may start sending extracted attachment bodies to
Rummager. Since this feature was never completed, we've not fully tested
it, and we're planning to remove it in its current form, we shouldn't
inadvertently enable it.